+++ pyteal/ast/__init__.py
@@ -156,7 +156,8 @@ from pyteal.ast.scratch import (
 )
 from pyteal.ast.scratchvar import DynamicScratchVar, ScratchVar
 from pyteal.ast.maybe import MaybeValue
-from pyteal.ast.multi import MultiValue, AddW
+from pyteal.ast.multi import MultiValue
+from pyteal.ast.multiexpr import AddW
 from pyteal.ast.opup import OpUp, OpUpMode
 from pyteal.ast.ecdsa import EcdsaCurve, EcdsaVerify, EcdsaDecompress, EcdsaRecover
 from pyteal.ast.router import (
